d9p1 done, am not happy
I was absolutely certain my approach to solving this, while a bit
crazy, would work, and was very confused as to why it would, so in
my frustration I asked an LLM what the frick is wrong with my code
and it was the smallest, dumbest thing. I somehow forgot the
'patterns' variable related to the buttons of the machines rather
than the lights, despite the fact I specifically use it to check
which buttons to press. So I was only checking the lowest 10 bits
of the patterns since the machines only have 10 lights.

I am deeply disappointed in myself. Both for missing something so
trivial and for resorting to using an LLM to help me with it.